% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_empty}
\alias{check_empty}
\title{Check if input vector is empty}
\usage{
check_empty(x)
}
\arguments{
\item{x}{Input vector of any length}
}
\value{
Numeric vector
}
\description{
Check if input vector is empty
}
\examples{
check_empty(12,4,5,6,7) # returns c(12,4,5,6,7)
try(check_empty()) # throws an error
}
