% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_na}
\alias{check_na}
\title{Check for missing values in the input vector(NA)}
\usage{
check_na(x)
}
\arguments{
\item{x}{A numeric vector of any length}
}
\value{
A numeric vector without NA and NaN values
}
\description{
Check for missing values in the input vector(NA)
}
\examples{
check_na(c(5,7,NA)) # triggers NA warning and removes NA
try(check_na(c(NA,NA))) # throws an error
}
