% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_single_value}
\alias{check_single_value}
\title{Checks the length of the vector}
\usage{
check_single_value(x)
}
\arguments{
\item{x}{Input vector of any length}
}
\value{
Numeric vector
}
\description{
Checks the length of the vector
}
\examples{
check_single_value(23,44,5,68)
try(check_single_value(5)) # displays a warning
}
