% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Descriptive2-package.R
\docType{package}
\name{Descriptive2-package}
\alias{Descriptive2}
\alias{Descriptive2-package}
\title{Descriptive2: Descriptive Statistics}
\description{
min, median, mean, hisogram, boxplot
}
\author{
\strong{Maintainer}: Damonzon Coulibaly \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
