% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_vector}
\alias{validate_vector}
\title{Combine functions above to handle edge cases (NA, single value, character)}
\usage{
validate_vector(x)
}
\arguments{
\item{x}{Input vector of any length}
}
\value{
Numeric vector
}
\description{
Combine functions above to handle edge cases (NA, single value, character)
}
\examples{
validate_vector(5,4,55,88,23, NA, 32) # returns c(5,4,55,88,23,32)
try(validate_vector(NA, "c", 6,2,3,1)) #throws error
}
