% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartiles.R
\name{calc_iqr}
\alias{calc_iqr}
\title{Calculate interquartile range (Q3-Q1) of input vector}
\usage{
calc_iqr(vals)
}
\arguments{
\item{vals}{A numeric vector of any length}
}
\value{
A single numeric value of interquartile range of the input values.
If a single value in the input, returns a warning and the value.
}
\description{
Calculate interquartile range (Q3-Q1) of input vector
}
\examples{
x <- c(10,20,30,40,50)
calc_iqr(x) # return 20
try(calc_iqr(c())) # empty vector
try(calc_iqr(c("a", "b", 5))) # character vector
try(calc_iqr(c(NA,NA,NA))) # removes NA and vector is empty
}
