% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_tendency.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate the mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The first quartile.
}
\description{
Calculates the mean of a numeric vector.
Missing values (\code{NA} and \code{NaN}) are removed.
An error is shown for empty, all missing, or single value inputs.
}
\examples{
x <- c(1,2,3,4,5)
calc_mean(x) # Returns 3

calc_mean(c(1, NA, 3)) # Returns 2

try(calc_mean(numeric(0))) # Shows an error
try(calc_mean(c(NA, NA))) # Shows an error
try(calc_mean(5)) # Shows an error

}
