% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_tendency_calc.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate mean to 2 dp}
\usage{
calc_mean(vals)
}
\arguments{
\item{vals}{A numeric vector of any length}
}
\value{
Mean value of a numeric vector
}
\description{
Calculate mean to 2 dp
}
\examples{
x < - c(10,20,30,40,50)
calc_mean(x) # return 30
# Edge cases - return error
try(calc_mean(c())) # empty vector
try(calc_mean(c("a", "b", 5))) # character vector
try(calc_mean(c(NA,NA,NA))) # removes NA and vector is empty
# Throws a warning, input length equals 1
try(calc_mean(c(4))) # return 4
}
