% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate Arithmetic Mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing the arithmetic mean.
}
\description{
Computes the arithmetic mean of a numeric vector after removing NA values.
}
\details{
\itemize{
\item NA values are silently removed before computation.
\item An error is raised if \code{x} is not numeric, is empty, or contains
only NA values.
}
}
\examples{
calc_mean(c(1, 2, 3, 4, 5))
# [1] 3

calc_mean(c(1, NA, 3))
# [1] 2

calc_mean(c(1, 2, 2, 3, 4, 5, 5, 5, 6, 10))
# [1] 4.3

}
\seealso{
Other descriptive statistics: 
\code{\link{calc_iqr}()},
\code{\link{calc_median}()},
\code{\link{calc_mode}()},
\code{\link{calc_q1}()},
\code{\link{calc_q3}()}
}
\concept{descriptive statistics}
