% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate Arithmetic Mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric scalar representing the mean.
}
\description{
Computes the arithmetic mean of a numeric vector. Automatically filters out missing (NA) values.
Throws an error if the input vector is empty, NULL, non-numeric, or contains only NA entries.
}
\examples{
data <- c(1, 2, 2, 3, 4, 5, 5, 5, 6, 10)
calc_mean(data)
calc_mean(c(1, 2, NA, 4))
}
