% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_tendency_calc.R
\name{calc_median}
\alias{calc_median}
\title{Calculate median to 2 dp}
\usage{
calc_median(vals)
}
\arguments{
\item{vals}{A numeric vector of any length}
}
\value{
Median value of the numeric vector
}
\description{
Calculate median to 2 dp
}
\examples{
calc_median(c(1, 2, 3, 4, 5)) # return 3
calc_median(c(1, 2, 3, 4)) # return 2.5
# Edge cases - return error
try(calc_median(c())) # empty vector
try(calc_median(c("a", "b", 5))) # character vector
try(calc_median(c(NA,NA,NA))) # removes NA and vector is empty
# Throws a warning, input length equals 1
try(calc_median(c(4))) # return 4
}
