% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_q3.R
\name{calc_q3}
\alias{calc_q3}
\title{Calculate third quartile}
\usage{
calc_q3(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The third quartile of `x`. Returns `NA` for an empty vector
  or when all values are missing.
}
\description{
Calculates the third quartile (Q3) of a numeric vector.
Missing values and `NaN` values are excluded from the calculation.
}
\examples{
calc_q3(c(1, 2, 3, 4, 5))
calc_q3(c(1, 2, NA, 4))

}
