% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartiles.R
\name{calc_q3}
\alias{calc_q3}
\title{Calculate third quartile (75th percentile) of input values}
\usage{
calc_q3(vals)
}
\arguments{
\item{vals}{A numeric vector of any length}
}
\value{
A single numeric value that cuts of the first 75% of the data sorted in ascending order.
If a single value in the input, returns a warning and the value.
}
\description{
Calculate third quartile (75th percentile) of input values
}
\examples{
x <- c(10,20,30,40,50)
calc_q3(x) # return 40
try(calc_q3(c())) # throws error, empty vector
try(calc_q3(c("a", "b", 5))) # return error character vector
try(calc_q3(c(NA,NA,NA))) # removes NA and vector is empty, throws error
}
