% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_numeric}
\alias{check_numeric}
\title{Check for non-numeric (character, Boolean, infinite) values in the input vector}
\usage{
check_numeric(x)
}
\arguments{
\item{x}{A numeric vector of any length}
}
\value{
A numeric vector
}
\description{
Check for non-numeric (character, Boolean, infinite) values in the input vector
}
\examples{
check_numeric(5,5,6,7) # returns c(5,5,6,7)
try(check_numeric(3,4,5,M,N)) # throws an error
try(check_numeric(Inf,-Inf) # throws an error
try(check_numeric(TRUE,TRUE,FALSE) # throws an error and avoids R conversion to 0 and 1
}
